Add support for `deprecation_list` in Add endpoint
Refers to CLOUDDST-3835
